Rename wasi-motor to wasi
