+++ package.json
@@ -37,7 +37,7 @@
   },
   "devDependencies": {
     "husky": "^7.0.4",
-    "lerna": "^4.0.0",
+    "lerna": "^7.0.0",
     "lint-staged": "^11.0.0",
     "prettier": "^2.3.1"
   }
