+++ packages/runtime/package.json
@@ -61,7 +61,7 @@
     "xterm": "^4.12.0",
     "xterm-addon-fit": "^0.5.0",
     "xterm-addon-web-links": "^0.6.0",
-    "@runno/wasi-motor": "^0.3.0",
+    "@runno/wasi": "^0.4.0",
     "@runno/host": "^0.3.0"
   },
   "gitHead": "812d21ec7c0c5d57c0274caaf7ff21c79ca79c1f"
