+++ packages/wasi/lib/main.ts
