+++ packages/wasi/lib/wasi/snapshot-preview1.ts
