+++ packages/wasi/lib/wasi/wasi-context.ts
