+++ packages/wasi/lib/worker/wasi-host.ts
