+++ packages/wasi/lib/worker/wasi-worker.ts
