+++ packages/wasi/programs/src/bin/cat.rs
