+++ packages/wasi/programs/src/bin/demo.rs
