+++ packages/wasi/programs/src/bin/echo.rs
