+++ packages/wasi/programs/src/bin/hello-world-stderr.rs
