+++ packages/wasi/programs/src/bin/hello-world.rs
