+++ packages/wasi/programs/src/bin/ls.rs
