+++ packages/wasi/programs/src/bin/read-file.rs
