+++ packages/wasi/programs/src/bin/return-arg-count.rs
