+++ packages/wasi/programs/src/bin/single-arg-is-hello.rs
