+++ packages/wasi/programs/src/bin/three-args-foo-bar-baz.rs
