+++ packages/wasi/programs/src/bin/write-read.rs
