+++ packages/wasi/programs/src/bin/write.rs
