+++ packages/wasi/src/main.ts
