+++ packages/wasi/src/vite-env.d.ts
