+++ packages/wasi/tests/core.spec.ts
