+++ packages/wasi/tests/libc.spec.ts
