+++ packages/wasi/tests/stdio.spec.ts
