+++ packages/wasi/vite.config.js
