+++ packages/website/package.json
@@ -23,7 +23,7 @@
   "devDependencies": {
     "@runno/host": "^0.3.0",
     "@runno/runtime": "^0.3.1",
-    "@runno/wasi-motor": "^0.3.0",
+    "@runno/wasi": "^0.4.0",
     "@tailwindcss/typography": "^0.5.4",
     "@types/marked": "^4.0.7",
     "autoprefixer": "^10.4.0",
